+++ src/utils/math.h
@@ -7,52 +7,52 @@
 #include <vector>
 
 struct Tuple {
-    float x, y, z, w;
+    double x, y, z, w;
     
-    Tuple(float x, float y, float z, float w);
+    Tuple(double x, double y, double z, double w);
     Tuple();
     
     bool isPoint() const;
     bool isVector() const;
-    float magnitude() const;
+    double magnitude() const;
     Tuple normalize() const;
-    float dot(const Tuple& other) const;
+    double dot(const Tuple& other) const;
     
-    float& operator[](int i);
-    const float& operator[](int i) const;
+    double& operator[](int i);
+    const double& operator[](int i) const;
 
     bool operator==(const Tuple& other) const;
     Tuple operator+(const Tuple& other) const;
     Tuple operator-(const Tuple& other) const;
     Tuple operator*(const Tuple& other) const;
-    Tuple operator*(float scalar) const;
+    Tuple operator*(double scalar) const;
     //Tuple operator*(int scalar) const;
-    Tuple operator/(float scalar) const;
+    Tuple operator/(double scalar) const;
     Tuple operator-() const;
 
     Tuple cross(const Tuple& other) const;
     Tuple reflect(const Tuple& normal) const;
 };
 
-Tuple operator*(float scalar, const Tuple& tuple);
+Tuple operator*(double scalar, const Tuple& tuple);
 Tuple operator*(int scalar, const Tuple& tuple);
 std::ostream& operator<<(std::ostream& os, const Tuple& p);
 
-Tuple Point(float x, float y, float z);
-Tuple Vector(float x, float y, float z);
-Tuple Color(float r, float g, float b);
+Tuple Point(double x, double y, double z);
+Tuple Vector(double x, double y, double z);
+Tuple Color(double r, double g, double b);
 
 
 struct Matrix {
-    std::vector<std::vector<float>> v;
+    std::vector<std::vector<double>> v;
     int height, width;
 
     Matrix(int height, int width);
-    Matrix(std::vector<std::vector<float>>& v);
-    Matrix(std::initializer_list<std::vector<float>> list);
+    Matrix(std::vector<std::vector<double>>& v);
+    Matrix(std::initializer_list<std::vector<double>> list);
   
-    std::vector<float>& operator[](int i);
-    const std::vector<float>& operator[](int i ) const;
+    std::vector<double>& operator[](int i);
+    const std::vector<double>& operator[](int i ) const;
 
     bool operator==(const Matrix& other) const;
 
@@ -61,19 +61,19 @@ struct Matrix {
 
     Matrix T() const;
 
-    float det() const;
+    double det() const;
     Matrix submatrix(int row, int col) const;
-    float minor(int row, int col) const;
-    float cofactor(int row, int col) const;
+    double minor(int row, int col) const;
+    double cofactor(int row, int col) const;
     Matrix inverse() const;
 };
 
-Matrix Translation(float x, float y, float z);
-Matrix Scaling(float x, float y, float z);
-Matrix Rotation_x(float rads);
-Matrix Rotation_y(float rads);
-Matrix Rotation_z(float rads);
-Matrix Shearing(float xy, float xz, float yx, float yz, float zx, float zy);
+Matrix Translation(double x, double y, double z);
+Matrix Scaling(double x, double y, double z);
+Matrix Rotation_x(double rads);
+Matrix Rotation_y(double rads);
+Matrix Rotation_z(double rads);
+Matrix Shearing(double xy, double xz, double yx, double yz, double zx, double zy);
 Matrix ViewTransform(const Tuple& from, const Tuple& to, const Tuple& up);
 
 extern const Matrix IDENTITY_MATRIX;
